+++ src/App.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Menu from '../examples/radio/App.svelte'
+  import Menu from '../examples/checkboxGroup/App.svelte'
 </script>
 
 <Menu />